4DC - T2 Aeon Sniper
Overhaul the way, (Azraeel) was removing the SHIELDPIERCING Category.

Upon farther review, Balthazar informed me; That I could not remove a category via merge in the Blueprint.

With this information, we are not properly removing the Shield Piercing via the Blueprints.lua and this no longer requires us to hold the Script for the Proj as well.