Fully fix Snap Behavior for the Mongoose
This combines the two ranges into a singular maxradius of 34, the grenadelauncher was fighting for control of the torso with the gatling gun.

This caused the strange "snapping" you would see with the mongoose when it was kiting at max range, this would also cause the gatling gun to lock up sometimes and not fire at all.