Patch to fix t2 miasma and other misses
Looks like upstream introduced an odd
TrackProj function that is invoked
only when the target has "m" in the
AI name (e.g. only for m28).

This manifested in several odd bugs,
namely aeon artillery miasma projectiles
just flying over targets entirely,
IFF the targets were m28AI controlled.

e.g. for LOUD, everything works fine.

We eventually isolated the problem to
the "new" circa 7.10/11 introduction
of the TrackProj monitor invoked
on the projectile in
defaultweapons.lua.

This patch hooks the function
and turns it into a noop,
which allows projectiles to
travel unimpeded (as they
do with LOUD).