Nerf T1 MAA Meta
I'm giving this one a description because I know how people feel about it. I want to say that the T1 MAA Meta leaves a sour taste in my mouth. Any spam like this later game should not be possible and should not result in a effective trade efficiency like I was seeing vs T3 Gunships. This nerf essentially doubles the cost of T1 MAA, this is also a indirect buff to T1 Bomber Usage in the early game as well.