Dynamic Factory Cost
This adds Dynamic Factory cost specifically when entering the Tech & Layer for the first time.

This means you'll receive a 2x buildtime with a 35% Mass & 80% Energy Increase for the first factory of tech & layer that you upgrade.

This is stored and remembered for the specific brain that performed that upgrade allowing each player to face this increase when they do their first initial tech upgrade.

This avoids the issue of importing all the code for the HQ System of FAF but gives us a similar system with much cheaper performance case and maintenance cost.